requirements.txt
flask~=3.0.2
flask-smorest
python-dotenv~=1.0.0
marshmallow~=3.20.1
sqlalchemy~=2.0.23
flask-sqlalchemy
flask_jwt_extended
passlib~=1.7.4
flask-migrate
psycopg2
requests
werkzeug~=3.0.1
alembic~=1.12.1
plotly~=5.19.0
